fix(flat): restore missed PreGenesis guards in the ulong migration
CI surfaced State/JsonRpc/Synchronization test failures all funnelling through
FlatDbManager.AddSnapshot, which silently dropped every committed snapshot:

    if (endBlock.BlockNumber <= persistedStateId.BlockNumber) { ... return; }

With nothing persisted, persistedStateId is PreGenesis (ulong.MaxValue), so the
raw `<=` is always true and the snapshot is discarded — state for any block is
then "concurrently removed". This is a sign-assuming comparison the ulong
migration missed: it compiles fine (so the build and State.Flat.Test passed) but
needs the `persistedStateId != StateId.PreGenesis` guard master's #11937 used.

Fixes the same class of missed guard in two more spots:
- FlatDbManager.FlushCache: `persistedState.BlockNumber < 0` (dead for ulong) ->
  `persistedState == StateId.PreGenesis`.
- PersistedSnapshotLoader.ReconstructBloom: finding the oldest base From with a
  raw ulong `<` never picks a genesis-spanning PreGenesis From; compare signed.

A full audit of every `.BlockNumber` ordering comparison in Nethermind.State.Flat
confirms these were the only unguarded ones. Now green: State.Test 835, JsonRpc.Test
1879, Synchronization.Test 2440, State.Flat.Test 778 — all 0 failed.

+++ src/Nethermind/Nethermind.State.Flat/PersistedSnapshots/PersistedSnapshotLoader.cs
@@ -166,7 +166,9 @@ public sealed class PersistedSnapshotLoader(
         // The persisted tier sits on the committed base — the oldest loaded snapshot's From.
         StateId committed = entries[0].From;
         foreach (CatalogEntry e in entries)
-            if (e.From.BlockNumber < committed.BlockNumber) committed = e.From;
+            // Signed compare: a genesis-spanning From is PreGenesis (ulong.MaxValue) yet sorts BELOW
+            // genesis, so the raw ulong `<` would never pick it as the oldest base.
+            if ((long)e.From.BlockNumber < (long)committed.BlockNumber) committed = e.From;
         if (head == committed) return;
 
         // Widest-first chain from head down to the committed base; .InMemory is empty at reload.
